llm-client: settings pane replaces dialog
Settings is now a full pane that replaces the chat area instead of a
modal overlay. Same layout as the sidebar — click the gear icon to
enter settings, back arrow to return to chat.

Content organized into sections: Prompts (system + seed), Model
(temperature + context override), and Behavior (auto-summarize with
budget slider + dedup retry). "Saved" flash confirms on save.

Gear icon in sidebar highlights when settings is active. Chat rows
deselect visually when in settings view. Old settings dialog removed.
